scrapy_app/requirements.txt
Scrapy==2.5.0
readability-lxml==0.8.1
pymongo==3.12.0